// pneus_max_sge/app/lib/types.ts
export interface Pneu {
  id: number;
  marca: string;
  modelo: string;
  medida: string;
  largura: string;
  perfil: string;
  aro: number;
  indice_peso: string;
  indice_velocidade: string;
  tipo_construcao: string;
  tipo_terreno: string;
  desenho: string;
  preco: number;
  quantidade: number;
  createdAt: string;
}